physics and some input

+++ src/Physics.cpp
@@ -11,8 +11,8 @@ bool PhysicsSystem::init(){
 void PhysicsSystem::update(unsigned int delay) {
 	for(auto p : physics)
 	{
-		p.second.positionCom->x = p.second.positionCom->x + p.second.xv*delay;
-		p.second.positionCom->y = p.second.positionCom->y + p.second.yv*delay;
+		static_cast<int>(p.second.positionCom->x = static_cast<float>(p.second.positionCom->x) + p.second.xv*static_cast<float>(delay));
+		static_cast<int>(p.second.positionCom->y = static_cast<float>(p.second.positionCom->y) + p.second.yv*static_cast<float>(delay));
 	}
 }
 void PhysicsSystem::cleanup(){}
